Remove links to old domain
Mike gave up his domain in 2019 after over 10 years during which time it was
hosted on a series of servers at my house in North Pole, Alaska until the end
of 2017 when I moved most of the virtual hosts to an instance in my AWS virtual
private cloud.

The remaining link is now a secure link.
Add secure links to my top level domain names.